pyproject.toml
[project]
name = "academicops"
version = "0.1.8"
description = "Academic Operations Framework"
requires-python = ">=3.11"
dependencies = [
  # Core runtime
  "pyyaml>=6.0",
  "pydantic>=2.0",
  "fastmcp>=2.13.1",
  "filelock>=3.13.0",
  "click>=8.1.0",
  "rich>=13.0.0",
  # Skills dependencies
  "pdfminer-six>=20260107",
  "python-ntfy>=0.1.0",
]

[project.optional-dependencies]
# Visualization tools (streamlit dashboard, graph visualization)
viz = [
  "streamlit>=1.28.0",
  "streamlit-agraph>=0.0.45",
  "networkx>=3.6",
  "matplotlib>=3.10.8",
]
# Anthropic API for skills that call Claude directly
anthropic = [
  "anthropic>=0.75.0",
]

[project.scripts]
task = "scripts.task_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["lib", "skills", "tests", "scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  # Testing
  "pytest>=8.0.0",
  "pytest-xdist>=3.8.0",
  "pytest-cov>=7.0.0",
  "pytest-timeout>=2.4.0",
  # Linting & type checking
  "mypy>=1.19.1",
  "ruff>=0.14.10",
  "types-pyyaml>=6.0.12.20250915",
  # Git hooks
  "pre-commit>=4.5.1",
]

[tool.ruff]
# Exclude archived files from linting - they are legacy code kept for reference
exclude = ["archived/**"]

[tool.ruff.lint.per-file-ignores]
# E402: Module level import not at top of file
# These files need sys.path manipulation before imports for the hook system
"aops-core/hooks/*.py" = ["E402"]
"aops-core/mcp_servers/*.py" = ["E402"]
"aops-core/scripts/*.py" = ["E402"]
"tests/**/*.py" = ["E402"]
"scripts/*.py" = ["E402"]
"polecat/*.py" = ["E402"]
"lib/overwhelm/*.py" = ["E402"]

[tool.mypy]
# Exclude archived files from type checking - they are legacy code kept for reference
exclude = ["archived/"]

[tool.pytest.ini_options]
testpaths = ["tests", "aops-core/tests"]
pythonpath = [".", "aops-core", "aops-core/hooks", "aops-core/lib"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-n 4 -m 'not slow'"
markers = [
  "slow: marks tests as slow (run with '-m slow')",
  "integration: marks tests as integration tests",
  "metrics: observability tests that report metrics (run with '-m metrics')",
  "demo: demo tests showing real behavior (run with '-k demo -v -s -n 0')",
  "endtoend: tests requiring real API credentials (run with '-m endtoend')",
]
